pyproject.toml
[tool.poetry]
name = "stable-fluids"
version = "0.1.0"
description = ""
authors = ["Thom <littlethomas98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.2"
scipy = "^1.14.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
